refactor(log): first try to remove cyclic dependencies
