+++ src/MemLogger.js
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2020 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+const { assign } = Object;
+const FormattedLoggerBase = require('./FormattedLoggerBase.js');
+
+/**
+ * Logs messages to an in-memory buffer.
+ *
+ * Formatter can be anything; default just logs the messages as-is.
+ *
+ * @class
+ * @implements Logger
+ */
+class MemLogger extends FormattedLoggerBase {
+  /**
+   * An array that holds all the messages logged thus far.
+   * May be modified An array that holds all the messages logged thus far.
+   * May be modified.
+   * @memberOf MemLogger#
+   * @member {Array} buf
+   */
+
+  constructor(opts = {}) {
+    super(opts);
+    assign(this, { buf: [] });
+  }
+
+  _logImpl(payload) {
+    this.buf.push(payload);
+  }
+}
+
+module.exports = MemLogger;
